JB Build Week Pskills
Initial File Structure Completed - Left a blank component available for the Marketing page, not sure if this is something that was required, open to change. Began usage of reactstrap components, laid out prototypical data structure in readme file. Laid out groundwork for Redux file structure. Began some work on styled components and allowed for addition of authorization header in the AdminView component.