Further Component Planning
Incorporated minimal styling to get an idea of how the page will look before I finalize the plans for my components. Was able to fully implement styles from the selected template. Began work on constructing the login view for the Admin and Employer. The admin view will be variable depending on whether given admin user is associated with any data objects yet.