Fully Functional CRUD w/ Prisoner ID Link
Moved all authentication operations to the redux store. Upon logging in, user is directed to the WorkersMgmt component which has a CDM that assigns the entire login response object to localStorage. this is used to reobtain user ID and authToken upon page refresh. Auth token and user ID are both kept in redux store to facilitate CRUD operations.